feat(@deepnote/database-integrations): extract reusable integration loading
Move the integration loading pipeline (YAML parsing, .env secret
resolution, cloud-pull/fetch, and comment-preserving YAML/.env write-back)
out of @deepnote/cli into @deepnote/database-integrations so the CLI and
the VS Code extension can load and store integrations the same way.

- Root export is browser-safe; Node fs/process.env wrappers live behind a
  new @deepnote/database-integrations/node subpath. File-touching functions
  are split into a content-accepting core (parseIntegrations, parseDotEnv,
  parseIntegrationsDocument, ...) plus thin path wrappers
  (parseIntegrationsFile, readDotEnv/updateDotEnv, ...).
- @deepnote/cli integration modules become thin re-export shims; the `run`
  command and CLI behavior are unchanged. Integration tests moved into the
  package.
- Reuse @deepnote/blocks' strict parseYaml/decodeUtf8NoBom for parity.
- Reimplement .env parsing (no dotenv dependency) to keep the root
  browser-safe, since dotenv eagerly requires node:fs.
- MergeResult now reports skipped (invalid/unsupported) integrations so
  callers can surface them.

Bumps @deepnote/database-integrations to 1.5.0.
