+++ packages/cli/src/integrations/integrations-file-schemas.ts
@@ -1,30 +1,8 @@
-import { databaseIntegrationConfigSchema } from '@deepnote/database-integrations'
-import { z } from 'zod'
-
-/**
- * Schema for the integrations YAML file structure.
- * Uses loose validation to accept any array of objects,
- * allowing per-entry validation to report specific issues.
- *
- * Example file format:
- * ```yaml
- * integrations:
- *   - id: my-postgres
- *     name: My PostgreSQL
- *     type: pgsql
- *     metadata:
- *       host: localhost
- *       ...
- * ```
- */
-export const baseIntegrationsFileSchema = z.object({
-  integrations: z.array(z.record(z.unknown())).optional().default([]),
-})
-
-export type BaseIntegrationsFile = z.infer<typeof baseIntegrationsFileSchema>
-
-export const integrationsFileSchema = z.object({
-  integrations: z.array(databaseIntegrationConfigSchema),
-})
-
-export type IntegrationsFile = z.infer<typeof integrationsFileSchema>
+// Integration file schemas now live in @deepnote/database-integrations. Re-exported here
+// for backward compatibility.
+export {
+  type BaseIntegrationsFile,
+  baseIntegrationsFileSchema,
+  type IntegrationsFile,
+  integrationsFileSchema,
+} from '@deepnote/database-integrations'
