+++ packages/cli/src/utils/api.ts
@@ -1,12 +1,2 @@
-/**
- * Error thrown when API request fails.
- */
-export class ApiError extends Error {
-  readonly statusCode: number
-
-  constructor(statusCode: number, message: string) {
-    super(message)
-    this.name = 'ApiError'
-    this.statusCode = statusCode
-  }
-}
+// ApiError now lives in @deepnote/database-integrations. Re-exported for backward compatibility.
+export { ApiError } from '@deepnote/database-integrations'
