+++ packages/database-integrations/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@deepnote/database-integrations",
-  "version": "1.4.3",
+  "version": "1.5.0",
   "description": "Deepnote database integration definitions",
   "keywords": [],
   "repository": {
@@ -15,6 +15,11 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
+    },
+    "./node": {
+      "import": "./dist/node/index.js",
+      "require": "./dist/node/index.cjs",
+      "types": "./dist/node/index.d.ts"
     }
   },
   "main": "./dist/index.cjs",
@@ -27,11 +32,13 @@
     "package.json"
   ],
   "scripts": {
-    "build": "tsdown --format esm --format cjs --dts",
+    "build": "tsdown",
     "test": "vitest",
-    "watch": "tsdown --watch --format esm --format cjs --dts"
+    "watch": "tsdown --watch"
   },
   "dependencies": {
+    "@deepnote/blocks": "workspace:*",
+    "yaml": "^2.8.3",
     "zod": "3.25.76"
   },
   "publishConfig": {
