+++ packages/database-integrations/src/loading/api-error.ts
@@ -1,0 +1,12 @@
+/**
+ * Error thrown when a Deepnote API request fails.
+ */
+export class ApiError extends Error {
+  readonly statusCode: number
+
+  constructor(statusCode: number, message: string) {
+    super(message)
+    this.name = 'ApiError'
+    this.statusCode = statusCode
+  }
+}
