+++ packages/database-integrations/src/loading/fetch-integrations.ts
@@ -1,0 +1,82 @@
+import { z } from 'zod'
+import { ApiError } from './api-error'
+
+/**
+ * Schema for a single integration from the API.
+ */
+export const apiIntegrationSchema = z
+  .object({
+    id: z.string(),
+    name: z.string(),
+    type: z.string(),
+    metadata: z.unknown(),
+    is_public: z.boolean(),
+    created_at: z.string(),
+    updated_at: z.string(),
+    federated_auth_method: z.string().nullable(),
+  })
+  .passthrough()
+
+export type ApiIntegration = z.infer<typeof apiIntegrationSchema>
+
+/**
+ * Schema for the full API response.
+ */
+export const apiResponseSchema = z
+  .object({
+    integrations: z.array(apiIntegrationSchema),
+  })
+  .passthrough()
+
+export type ApiResponse = z.infer<typeof apiResponseSchema>
+
+/**
+ * Fetch integrations from the Deepnote API.
+ *
+ * Uses the global `fetch`, so it works in Node (18+) and browser-like environments alike.
+ *
+ * @param baseUrl - The base URL of the Deepnote API
+ * @param token - The authentication token
+ * @param integrationIds - Optional list of integration IDs to fetch. When provided, only these integrations are returned.
+ * @returns Array of integrations from the API
+ * @throws ApiError if the request fails
+ */
+export async function fetchIntegrations(
+  baseUrl: string,
+  token: string,
+  integrationIds?: string[]
+): Promise<ApiIntegration[]> {
+  const endpoint = new URL(`${baseUrl}/v2/integrations`)
+  endpoint.searchParams.set('includeMetadata', 'true')
+  if (integrationIds && integrationIds.length > 0) {
+    endpoint.searchParams.set('integrationIds', integrationIds.join(','))
+  }
+  const url = endpoint.toString()
+
+  const response = await fetch(url, {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    },
+  })
+
+  if (!response.ok) {
+    if (response.status === 401) {
+      throw new ApiError(401, 'Authentication failed. Please check your API token.')
+    }
+    if (response.status === 403) {
+      throw new ApiError(403, 'Access denied. You may not have permission to access integrations.')
+    }
+    throw new ApiError(response.status, `API request failed with status ${response.status}: ${response.statusText}`)
+  }
+
+  const json = await response.json()
+
+  const parseResult = apiResponseSchema.safeParse(json)
+  if (!parseResult.success) {
+    throw new Error(`Invalid API response: ${parseResult.error.message}`)
+  }
+
+  return parseResult.data.integrations
+}
