+++ tsconfig.json
@@ -11,6 +11,7 @@
     "noImplicitAny": true,
     "noImplicitThis": true,
     "paths": {
+      "@deepnote/database-integrations/node": ["packages/database-integrations/src/node/index.ts"],
       "@deepnote/*": ["packages/*/src/index.ts"],
       "@test-fixtures/*": ["test-fixtures/*"]
     },
